Configurable changes to change database in local from PgSQL to Oracle 19c
