+++ AZSIS-Nexus-Pigeon-Service/src/main/resources/logback-spring.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <configuration>
 
-    <property name="APP_NAME" value="${spring.application.name:-DSC-Translation-Service}"/>
+    <property name="APP_NAME" value="${spring.application.name:-AZSIS-Nexus-Pigeon-Service}"/>
     <property name="LOG_DIR" value="logs/${APP_NAME}"/>
     <property name="LOG_LEVEL" value="${LOG_LEVEL:INFO}"/>
