--- AZSIS-Nexus-Pigeon-Service/src/test/java/com/daimlertrucksasia/it/dsc/localization/DSCLocalizationApplicationTests.java
@@ -1,13 +1,0 @@
-package com.daimlertrucksasia.it.dsc.localization;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DSCLocalizationApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
