+++ AZSIS-Nexus-Pigeon-Service/src/test/java/com/daimlertrucksasia/it/dsc/localization/PigeonServiceTests.java
@@ -1,0 +1,13 @@
+package com.daimlertrucksasia.it.dsc.localization;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PigeonServiceTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
