+++ AZSIS-Nexus-Traffic-Manager/src/main/resources/application.yaml
@@ -1,5 +1,5 @@
 spring:
   application:
-    name: ${SPRING_APPLICATION_NAME:apigateway}
+    name: ${SPRING_APPLICATION_NAME:AZSIS-Nexus-Traffic-Manager}
   profiles:
     active: ${SPRING_PROFILES_ACTIVE:dev}
